server/requirements.txt
Flask
Flask-MongoKit
python-social-auth
flask-login
six
